Issue #295: Focus main container first when other views are missing.
We originally placed a dummy focus view to obtain focus, because we were in a
LinearLayout which must specify the views in order, but that didn't work well
because this view would be focused at unintuitive times, both for average users
and VoiceTalk users. This commit replaces that dummy view with the main
container, now in a RelativeLayout to ensure it's first.