Closes #295: Toggle cursor when VoiceView is toggled.
The behavior isn't perfect but it's now usable. In particular, we may want to
consider automatically entering review mode when relevant (when webview is
shown, when keyboard is present) but there's a lot of state to manage there
(what if the user previously disable review mode, we open the menu, and come
back to the webpage?).