Rm youtube reload hack when opening menu.
The reason youtube did not have spatial navigation is that the webview wasn't
being focused but now that the webView is at the top of the view hierarchy
(rather than the dummy view), this is no longer a problem.

Note: I only tested this on an emulator, not on device.