Closes #311: re-enable tracking protection.
I think request.isForMainFrame() is unnecessary because of this line in
UrlMatcher.matches:

        if (pageHost != null && pageHost.equals(resourceHost)) {
            return false;
        }

Which whitelists first party requests: this is assuming the `pageHost`, which
comes from `TrackingProtectionWebViewClient.currentPageUrl`, actually
represents the current page loading (and thus the main frame).

I noted that my Android TV emulator seems to block the same numbers of trackers
as Focus.