Closes #325 - review: controllerBackEvent -> browserBackControllerEvent.
We really want to track browser back events. There are two ways to go back:
through the controller and through the menu. We can either:
- Make the back button APIs consistent (browserBack(menu/back) )
- Make the menu APIs consistent (menuPress(back/...) )

Since the menus are actual colocated buttons, rather than the back buttons
which are actions, I felt it was more important to keep the menu actions
consistent. As such, I kept the controller back event separate.

---

I converted controllerBack -> browserBack because we're more concerned about
the actual action taken in the browser rather than how many times users have
pressed the physical buttons on the remote.

---

This probe is intentionally only fired when the browser can go back, i.e. not:
- When back is pressed when the sidebar is open over the browser
- When back is pressed on YouTube, which is now an ESC key