Issue #329: Add telemetry for global tracking protection state.
Users should set TP in onboarding so it'll either be TRUE or FALSE. If they
manage to avoid setting it (maybe back in onboarding), it will be NULL.

I don't really have time to think through future changes (e.g. if we make it a
preference per page) and how we should prepare the telemetry for that case:
I guess we'll just make a new probe or something.