Issue #403: Rm immersive mode code.
There are no system status bars on Fire TV so disabling them is no
longer important. Also, the screen seems to stay on when playing a video
in my local testing (I'm not sure why) so I don't think the requirement
to add flags to leave the screen on is necessary either.