Issue #414: Add IWebViewLifecycleManager to replace WebFragment.
WebFragment has two responsibilities:
- Manage the lifecycle of the web view
- Make locale callbacks on it

We replace both of these with this class, though the non-WebView
functionality of locale switching will be moved elsewhere.

I did not like WebFragment as an abstraction because the call order of
the method it overrides and requests to be abstract make the super and
sub class tightly coupled. Instead, this class requires more explicit
calls to use but is clearer by maintaining a single responsibility and
generally more maintainable by being composable, rather than inherited.