Issue #412: Add CursorKeyDispatcher.
This lets the ViewModel drop one responsibility simplifying its public
interface to high level methods (onSpecificDirPressed). This helps
understand the possible states and makes it easier to test (i.e. simpler
than mocking a KeyEvent).

It adds a little complexity to BrowserFrag now that there are three references
and lifecycles to maintain:
- View
- ViewModel
- KeyDispatcher