Issue #412: Move cursor updates from loading into Cursor.
Most of the cursor code is now encapsulated inside the controller. One
downside to this approach is that it's difficult to see all the changes
to happen to one event (e.g. what changes when accessibility state
changes?) but it's much easier to see all the changes that affect the
cursor state (as opposed to scanning around the BrowserFragment file and
all its lifecycle methods).