Closes #493: Set root container to focus descendants first.
This prevents focus from getting stuck on the root container, rather
than one of its children views.

It's not a perfect solution: when we back out of the soft keyboard, the
last home tile is focused rather than the url bar, but I think it's good
enough for the amount of time I invested in it.

This does not fix the Youtube reload hack.