Closes #276: Disable autocomplete when using remote app.
We can delete the IME reset code in deleteSurroundingText because it
doesn't seem to help on Fire TV. With the soft keyboard, it does not
appear to do anything. With the remote app, it makes backspacing
autocompleted results broken. I removed it, which normalizes the
backspacing.

I didn't follow the code in `afterTextChanged` closely enough to know
why the change is necessary but:
- If we don't, when entering text on the remote app, sometimes a small
section of it is deleted - somewhere, we're updating the text entry
behind the scenes
- The existing code sets a boolean on whether or not we want to use
autocomplete so it makes sense we'd disable that boolean

---

Here's the full reasoning for why we disable autocomplete rather than
implement:

**I decided that I'm going to try to disable autocomplete when the
remote is in use, rather than trying to put autocomplete on the app**:
- I don't have a solid understanding of how the IME interacts with the
Editor (our InlineAutocompleteEditText), which I think is important to
creating a proper solution
- It may not even be possible to add autocomplete to the app (the way
we're doing it): the remote seems to take several async operations to
update: 1) type nothing, 2) delete all text, 3) type the original text +
the new character. In the mean time, it keeps its own record of what the
on-screen text should be. If we update the text on the device, the
remote may never be able to synchronize properly
- There are many edge cases to worry about: entering text with(out)
autocomplete, deleting text (and updating autocomplete), moving the
cursor manually, text selection, etc.
- If we implement a custom keyboard as Tif wants, we're going to remove
inline autocomplete anyway so this work will be for nothing