Closes #557: Use Delegates.observable instead of setter.
observable is safer because you can't accidentally change the value
being set and it's more readable.

observable does not work in all cases - using observable in
CheckableImageButton throws a NPE. I think this is because the property
is referenced by onCreateDrawableState, which is referenced by the super
constructor and these properties can't be referenced before the
constructor has returned. I deduced this from this
discussion: https://discuss.kotlinlang.org/t/npe-with-delegate/1808/4