Issue #511: Fix tests when running on the command line.
The tests passed in the IDE before because they working dir is the
project root but failed on the command line because the working dir is
the module, hence my hard-coding didn't work. This still hard-codes but
does it relative to the working directory.

Sorry to have undone some code I already wrote but it'd be hard to merge
with the rename.