Closes #511: Run tests with UTF-8 in TaskCluster.
The tests passed locally after this patch series but failed on TC. I discovered
running the tests in the docker image with LANG=en_US.UTF-8 fixed the issue,
making me think that the docker image forces applications into a non-UTF-8 mode.
Since we added tests that rely on UTF-8, the tests fail. For more details, see
https://github.com/mozilla-mobile/firefox-tv/issues/594

This isn't the ideal solution - it should probably be inherit to any
shell created in the image - but it should be good enough for now.