Closes #662: Remove unused app icons.
In practice, Amazon pulls all assets from our store page so none of
these are used. In development:
- app_banner is used on Android TV emulators
- app_launcher is used on Fire TV dev builds (including release and
beta)

One side effect of this change is you can't tell the difference between
beta/release/debug builds on the Fire TV anymore, but now we're not
shipping as many unused assets.