Issue #472: Use scrollBy for Cursor screen scrolling.
flingScroll started to look awful when raised to 60 FPS so I switched to
scrollBy, which was apparently available on View this whole time. It
should be smoother: we're no longer creating fling gestures with the
ScrollView each time the frame updates. This patch attempts to
keep scroll speed the same.