Issue #507: Move SessionManager intent handling code to IntentValidator.
THE TESTS WILL FAIL UNTIL THE NEXT COMMIT: I do this for ease of review.

The SessionManager doesn't need to handle this code and in the next
commit, we'll modify it so that it's no longer creating sessions.

The moved code should be identical in functionality to the existing code.