Closes #507: use showBrowserScreenForUrl on valid browser intents.
We used to create a new session every time we receive an intent, which
would add the existing BrowserFragment/WebView to the back and stack
causing redundant WebViews and resource usage.

At the end of this patch series, when we receive an intent, the behavior is:
- If BrowserFragment is visible, we re-use it
- If BF is not visible, we'll create a new Session, which creates a new
BrowserFragment. This could cause multiple WebViews if BF was in the
back stack but this is unlikely with the current code (only when
Settings is shown)
- Since we create and destroy many BF/WebView instances (e.g. home is
shown so we destroy and when a site is clicked, we recreate it), we're
relying on GC to ensure there is only one WebView instance at a time.
Curiously, force GC in the debugger does not seem to always remove inactive
WebView instances: either there's a bug in force GC or someone is
holding a reference to it and it only gets cleaned up after several GC.
It would likely be most resource efficient (assuming WebView is resource
intense even when it's inactive) to  re-use a single BrowserFragment/WebView.

This change uncovered a lot of cruft in the way we handle sessions (i.e.
we can create more than one), the browser fragment and session (it was
really hard to load a URL from outside the browserFragment), state
restoration, and intent handling.