Issue #774: Add focused border to Pocket mega tile.
This implementation makes design concessions:
- We don't do multiple rings in the border #837
- The dark blue border in the mock is 4dp inside the box, 4dp outside.
For simplicity, we do 4dp inside the box.

---

After this PR series, we still need to resolve any UX nits Amin may have
and figure out how to size the images (they're not to spec because their
aspect ratios are different than the specs for the views: e.g. should we crop?)