Issue #887: Selectively requestFocus on update in overlay.
The previous implementation broke dpad navigation when the ScrollView
was added: it caused two views to be focused, the url bar and whatever was
clicked. This implementation only requests focus when necessary and
seems to work correctly.

This previous implementation does not have comments or a great commit
message but it seems to have been added to fix the problem of losing focus
when the view we had focused was disable (e.g. back when we hit the end of the
back stack). It was implmented in 8b8690dd074f317b3da7a8054e31a6ffe65d0bff