Issue #940: Give BrowserFrag a reference to MediaSession.
This reference will be used to add various callbacks for the WebView
lifecycle.

This is the recommended way to get a reference to an Activity: have the
Activity implement an interface (to decouple one Activity from another)
and use the global activity reference. See the docs:
  https://developer.android.com/guide/components/fragments#EventCallbacks

If an object is parcelable, they can be added as an object too but our
object needs to have a single reference and is not parcelable.