Closes #983: Pause video immediately after play if paused issued while paused.
See comments for details.

This solution isn't perfect: as expressed in the comments, there is an
opportunity for us to pause a video that was intentionally played.

One possible alternative would be to pause all videos when voice
commands are issued and resume them when the voice command completes
but, if my theory is correct that audio focus changes are what cause
the videos to pause, this is called before any callbacks we have access
to and it's not possible to implement.