Issue #1204: Remove properties already set in DefaultSettings.
Components implemented their own Settings abstraction which overwrites
the values set by our EngineView.setupForApp extension function.

Our existing code already implements the new abstraction: we extend
DefaultSettings and pass it into our SystemEngine in Components.kt.

This patch is functionally a no-op but is used as book-keeping: it removes
values from our extension function that are already defined as
DefaultSettings to be the same value as the one we provide. The values I'm
comparing against are defined here:
https://github.com/mozilla-mobile/android-components/blob/1d94310f89f5d6c93797c71eb809ab3b84fb74b7/components/concept/engine/src/main/java/mozilla/components/concept/engine/Settings.kt#L106