Issue #1204: Remove properties SystemEngineSettings applies globally.
The components Settings abstraction has values we can override - through
Settings - and values we can't override.

This patch is functionally a no-op and is used for book-keeping: we
remove the properties from our extension function that we can't override
in Settings, which I verified are the same values we set locally. The
values I'm comparing against are defined here:
https://github.com/mozilla-mobile/android-components/blob/1d94310f89f5d6c93797c71eb809ab3b84fb74b7/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineSession.kt#L210