Closes #1392: removed disableDesktopModeWhenHostChanges
This method added a LiveData#observeForever call within a ViewModel, which has a constrained lifecycle. This created a leak, and was generally a bad practice.

State will now be checked whenever it flows through the mapping function in ToolbarViewModel.state. This causes a side effect from within a map call. This is a bad practice from a FP perspective but... at least it doesn't leak.