Closes #1392: updates ClearDataTest
This updates the test to current behavior, but still fails. After data is cleared and a new URL is navigated to, SessionLoadedIdlingResource blocks the test and times out. This occurs on master (if its test is updated to include that condition) as well, and does not occur when manually testing the app.  This appears to be related to #1360