Issue #1651: remove outdated comment in PocketVideoRepo / post value
This comment was out of date, we are not currently treating the value as nonnull.

setValue must be called from the main thread, while postValue can be called from anywhere.  This means that initializing the PocketVideoRepo from Espresso tests before the Application has been created crashes if setValue is used in init.