Closes #1394: Fixes problem where viewing multiple Pocket videos broke YouTube
Bug:
- Open Pocket screen
- Open Pocket video
- Press down on D-pad
- Open Overlay
- Open Pocket screen
- Open Pocket video

(if the video did not have an ad) D-pad will not work

RCA:
Focus is on the wrong DOM element. Theory is that YouTube is programatically requesting focus relatively early in the page lifecycle, and because commit is asychronous, it is being called before said DOM elements are ready to be focused.  Switching to commitNow runs this code synchronously, meaning that our fragment is ready to take focus by the time loadUrl is called.