Closes #1691: fixes inconsistent state assertion
RCA:
When clearing data, the displayed screen was changed to NavigationOverlayFragment, however the ScreenController did not actually update its currentActiveScreen value. This caused a disconnect between expected and actual current state. At this time, no user impact has been noticed, but there could be strange transaction bugs related to this.

Fix:
Setting up fragments for a new session is guaranteed to display the NavigationOverlayFragment, so instead of asserting that state, the method now sets it.

Regressions:
To avoid regressions, we should work out a system for assertions that crash on debug and report to Sentry silently on release, then periodically assert app state throughout ScreenController to guarantee that no similar situations exist in the app.  Opened https://github.com/mozilla-mobile/firefox-tv/issues/1693