Closes #1282: Remove unused strings in files containing youtube_casting_onboarding_toast.
afaik, we want to keep this string for when it will be used. These
string files had the string that was autoremoved so I needed to edit the
commits by hand (via `git add -p`) to ensure the string was kept. After
I completed the removal, I verified the strings were still there by
counting the number of matches in grep before and after (8).

Note that I also had to remove string localization comments in the
english strings by hand.