Issue #1664: Move toolbar UI code to ToolbarUiController.
The NavigationOverlayFragment is starting to get unwieldly: it's 550+
lines and it's hard to figure out which code is related to the toolbar
and which code is related to the bottom navigation panel. Since they're
being split, now is an opportune time to separate their file contents.
Moving the toolbar code out:
- Reduces NavOverlayFrag to ~430 lines
- Makes the dependencies between the toolbar and bottom navigation panel
more obvious
- Makes the dependencies between the toolbar and the framework more
obvious
- Makes the shared code more obvious

No functionality was changed and no code was cleaned up: I essentially
just moved functions, updated references, and added the necessary
parameters.

This refactor isn't perfect - e.g. should the NavOverlayFragment really
always handle `onNavigationEvent`? - but is intended to make things easier
to change later.