Closes #1748: Upgrade test orchestrator to latest.
Speculative fix: this may also fix #1755. These issues are intermittent
and not reproducible by everyone on every device. No-Jun removed the
test orchestrator and he was no longer able to reproduce his tests failing
for the same reason. I changed the test orchestrator and I am unable to
reproduce, however, I'm unable to reproduce when I use the old version
of test orchestrator either. Upgrading the version seems like a safe
move to try to fix the issues we're having.

The production code doesn't seem to be a problem: the loading state
(which the session idling resource depends on) updates correctly (false,
true, false) when running the application but does not update correctly
when running the test (false, true).