Issue #750, Issue #1628: Move Sentry init to Application.
We verified with the Sentry team that this is an acceptable way to
initialize Sentry.

This change is beneficial because:
- It fixes a memory leak
- It allows all Activites to get crash reporting
- We get crash reporting earlier in the init process

It's concerning because:
- we're modifying crash reporting code
- Sentry is slow to start up so this may further affect startup time

This change could be better if:
- We could initialize Sentry before the ServiceLocator is initialized
and before the PreferenceManager default values get set

I verified this change was working correctly by adding code to crash the
app (throw when M key is pressed) and ensuring the events were received
by the server. I also tested events are not received by the server when
data collection is disabled.