Issue #1628: Use new CursorViewModel.isEnabled in application.
There was a memory leak where the accessibility listener would not
release the reference to the WebRenderFragment, according to LeakCanary.
The trace was:
- AccessibilityManager.sInstance
- AccessibilityManager.mTouchExplorationStateChangeListeners
- ArrayMap.mArray
- Object[].[2]
- CursorController.webRenderFragment
- WebRenderFragment

I'm not sure why: it's possible the fragment's context was null so we
were unable to remove the listener. Rather than diagnose the issue and
fix it directly, it seemed better to remove the possibility of a
memory leak altogether and move towards our ideal design by
refactoring the code to use MVVM via the FrameworkRepo: the Framework
repo attaches the accessibilityManager to the Application Context so
there is no opportunity for leaks.

After the fix, I no longer see this leak in LeakCanary's results.

---

I could divide this into better commits but I didn't think it was worth
the time. This patch does a few things:
- Removes isEnabled functionality from CursorController
  - Removes accessibilityManager listener set-up/tear down
  - Removes current url/isLoading listener set-up/tear down
  - Removes callback for screen change events (and call sites)
  - Removes setEnabledForCurrentState logic (in favor of CursorViewModel)
- Uses CursorViewModel in fragment
- Creates CursorController with static function to call for onCreateView too