Closes #1801: fixes WebViewCacheTest
RCA:
WebViewCache previously instantiated its SystemEngineView by hand, but this was updated in 7f57f23 to delegate to WebRenderComponents.engine.  This object, however, requires a valid user agent before it can be instantiated, and none is provided in a test environment.  This can be fixed by supplying a fake user agent in the @Before block of the test.