No issue: disable codecov status checks.
Since codecov posts the coverage changes as comments in the PR, we no
longer need status checks (and thus their failures) to tell us the
coverage diffs. We could reconfigure such that the checks exist but
never fail but their docs are unintuitive so this seemed easier.

See this PR on FFES for what this looks like:
https://github.com/mozilla-mobile/firefox-echo-show/pull/232