WIP: Bitbar Cloud UI Testing
This PR integrates a Gradle plugin for supporting Bitbar Cloud UI testing on PRs. Configuration has to be defined in a Gradle task. Currently it uses a preset “Firefox for Fire TV” project defined on Bitbar using a predefined group of devices under a group name “Amazon”. We can customize the group selection through the front-end. Test run names are just a build + date (we can probably improve on that, just temporarily  using this approach).