Closes #1830: fixes YouTube focus when overlay is open
RCA:
There is a WebView bug that, when the WV loses and regains focus, causes the incorrect DOM element to receive focus (https://bugs.chromium.org/p/chromium/issues/detail?id=826577). Our workaround for this caches the currently focused view on overlay open, then focuses it on overlay close. When we open the overlay during a loading screen, we cache the body element. The overlay is then closed, and the body element focused, rather than the view on which YouTube has programmatically requested focus.

Fix:
Instead of caching focus on overlay open, this now caches it just before overlay close.  This allows us to use the most recently focused DOM element.