Issue #1584: Handle back out of YouTube from Pocket
This commit refactors the previous implementation of backing out of YouTube, which only handled the case of visiting from the YouTube tile.
Now, instead of searching for the CDN url, we save the index of the url that was visited previous to going to YT, then go back to that url on back press.