Closes #1830: set currentActiveScreen before changing fragments
Note that, due to the asynchronous nature of most of our fragment transactions, in most cases this value was set before any transaction took place anyway. This was not true for a the few cases that used commitNow, which executes synchronously. This change makes this behavior consistent, always updating currentActiveScreen just before the change is made, and makes that more clear to anyone reading the code.

Current consumers of currentActiveScreen were scrutinized, and none seem to conflict with this change.