Issue #1554: updated PocketVideoFragment to consume Pocket data via Rx
Moved the PocketViewModel.state subscription from onCreateView to onStart. This meant moving some other setup there as well. None of these moves look risky, and manual testing is behaving as expected.

This move was made to block against a very unlikely but possible future edge case. Committing a fragment transaction after onStop has been called will cause a crash, so if any code that could lead to a transaction was ever added to this subscription, cleaning it up in onDestroyView would present a narrow window where a crash would be possible.

AFAICT, this narrow edge case was present with LiveData too.  +1 point for manual lifecycle handling!