Issue #1912: Add ObservableMutableSet.
There is a bug in the current Cursor implementation where, when the
directionKeysPressed Set changes, we need to update the state of an
Observable but we forgot to in one place. This class will be used to
ensure this class of errors never happens again.

Due to time constraints, there are no tests.