Closes #2110: fixes FFVMProvidersTest on GV builds
The class tested here on master was removed. It was replaced in an earlier commit with the OverlayVM, which requires a SessionRepo.  This depends upon WebRenderComponents, which for some reason NPEs on a lateinit property on GV builds.  Since this is not what the test is meant to verify, rather than going down the rabbit hole of figuring out how to set up WebRenderComponents on a GV unit test, I've just switched the VM for a different one.