Issue #1652: break glean init into helper fun
We're mixing levels of abstraction in onCreate.  Most of the work done here is at a very high, declarative level (init something, pass some requirements), so mixing in the nitty gritty of observing data collection state makes the method harder to follow.  Solved by breaking it into a helper function.

This also makes it clear that the lines related to setUploadEnabled and initialize are more related to each other than they are Sentry and TelemetryIntegration init.