Issue #2240: Fix telemetry.md for broken menu open/close events.
"Menu opened by menu key" was originally created to represent them but,
as someone who was not on the project, I'm concerned between its original
implementation and when it completely broke, it may not have represented that
value (e.g. instead, representing menu is opened). However, I don't know if
it's worth conveying that concern in the telemetry docs because it may
be unfounded and we're unlikely to look back on that data.