Issue #2354: fix crash upon deleting last pinned tile
RCA:
The focus repo would change focus to Pocket any time 1) a home tile was focused, and 2) pinned tiles was empty. Pocket tiles are now considered home tiles, so this led to an infinite loop. With all channel tiles sharing the same ID, we now treat this as lost focus and move it to the URL nav bar.  A BETTER SOLUTION would be to move focus to the nearest channel, but this can unblock us for release.