Closes #2179: fixes PinTileTests
This block of Espresso is saying "find a view with an ID of home_tile, who is the first child of a view with an ID of pinned_tiles_channel.  Make sure that the parent is the second child of a LinearLayout."  A new view was added to that LL, so this needed to be updated to look for its 3rd child.

We desperately need to rewrite this test following the robot pattern, but are blocked until Fastlane screenshots work with Kotlin.