Closes #2539: fix tests that hang on CI
Root cause analysis:

If the code under test changes threads, use `RxPlugins` & `RxAndroidPlugins` to make sure it works properly during testing.  See "How can we avoid this in the future?" for details.

```kotlin
// PocketVideoRepo
open val feedState = _feedState.hide()
    .observeOn(AndroidSchedulers.mainThread())
    .distinctUntilChanged() // avoid churn because we may retrieve similar results in onStart.

```
The `observeOn` jumps on to the Android main thread.  Normally this would crash in a test and then be fixed, but Robolectric is allowing it to run.  For some reason, probably buried deep inside Robolectric, it’s working differently on CI than it is locally.  I was not able to work out why.

The fix is to tell the test to run this code on its default thread by including this in the test class.
```kotlin
// PocketVideoRepoTest
companion object {
    @JvmStatic
    @BeforeClass
    fun beforeClass() {
        RxAndroidPlugins.setInitMainThreadSchedulerHandler { Schedulers.trampoline() }
        RxAndroidPlugins.setMainThreadSchedulerHandler { Schedulers.trampoline() }
    }
}

```

_Normally_ only the first of those two calls should be necessary [(see this article for an explanation of the difference)](https://medium.com/@peter.tackage/overriding-rxandroid-schedulers-in-rxjava-2-5561b3d14212), but for whatever reason that is not working here.  Possibly some strange interaction with Robolectric?

If your code under test changes threads, make sure that they’re being properly handled in the test file, either with a `TestScheduler` (to mock the passage of time) or `trampoline` (to force everything on to the same thread).  `setInit*SchedulerHandler` _should_ always be enough, but it wasn't here.  Some people always set both versions just in case, **and I think we should probably default to doing that** for the sake of simplicity.

- What is the difference is between local and CI environments?
- Why is `setInitMainThreadSchedulerHandler` insufficient here?