Closes #2619: fixes bug that caused toolbar to be unresponsive to changes
The ToolbarVM was querying state from SessionRepo.legacyState in order to determine what the new version should be.  Previous commits in this branch removed the last observer of SessionRepo.legacyState and, interestingly, it seems that when the value of an unobserved LiveData is queried, it always returns null.  This caused toolbar interactions to break.  The fix here was to remove legacyState entirely, and migrate to Rx.