Issue #2491: Increase custom toast duration; move unpin toast to custom layout.
All toasts should use the custom layout. This toast needs to be duration
long. It would take a little time to change the custom toast code to
support variable durations but increasing it for all toasts has a minimal
impact on users so I don't think it's worth the time investment to fix
it.