Closes #2714: Replace unused app_banner with ic_launcher.
This is used by Android TV as the app tile.

This is unused by Fire TV:
- The production app tile is distributed via the Amazon dashboard
- The debug app tile uses `android:icon`, which is `ic_launcher`

Since we don't ship to Android TV, we just replace the app icon with
something smaller but still representative of Firefox (for emulator
builds) and that is already used as the app icon in debug Fire TV
builds: `ic_launcher`.