Issue #2662: rebase TrackingProtectionPolicy config on latest a-c.
We keep our settings functionally the same while working around a-c
bugs: these are commented in the code.

One minor change we made is we added TrackingCategory.TEST: according to the
code comments, it doesn't do anything and it's a part of the RECOMMENDED
configuration, which is ideally what we'd use but can't due to an a-c
bug.

I tested TP was working correctly by visiting rottentomatoes.com and
ensuring the banner ad appeared or disappeared as expected.