Closes #2493: fixed reauthentication ping
It was noticed that we had no good guarantee about when our settings ping would be sent (especially after we eventually migrate to Glean), meaning we could miss important information. For example, if sent on startup, 0% of users would be authenticated because the call would not yet be completed. Instead we chose to send all FxA state changes, but to narrow them down to 'Needs Reauth == true/false', to be as lean as possible.

We are also debouncing this ping, to cut down on redundant information. See FxaRepo#setupTelemetry for details