Fixes #2854: freeze when sending tab while on FxA profile
RCA:
Previously, we were adding the browser fragment and moving focus to it as expected. However, we were not removing the settings fragment, which covered the content. This left the user unable to do anything.

In the future, we may need to reevaluate our ScreenControllerStateMachine. Pushes can occur at any point of the application lifecycle, which complicates the defined edges between our points.