Issue #2518 - post: Call super.onNewIntent in MainActivity.
The linter threw an error that we were not calling through to super when
it was expected so I added the call. To be honest, I'm not sure what the
effects of this call will be. The super method javadoc says,

> If you are handling new intents and may be making changes to the
> fragment state, you want to be sure to call through to the super-class
> here first.  Otherwise... [under some conditions]... will throw
> IllegalStateException...

Under these conditions, calling through to super will (probably) prevent
us from seeing this IllegalStateException. However, I'm concerned about
what the other side effects of adding this call may be. However, I don't
know of a good way of finding out besides adding it and seeing if there
are side effects and this approach seems better than suppressing the
linter warning.